modules/cloud_account/aws_provider/1.0/requirements.txt
boto3
google-cloud-secret-manager
